feature/era-glonass-integration-r2: VIN synchronizer:
moved `unmarkAsHandled` to own sub-module
